add friendship status in service, add security controls in backend friendishipController

+++ server/src/controllers/friendship.controller.ts
@@ -50,19 +50,29 @@ export const searchFriends = (req: Request, res: Response) => {
 
   try {
     const { q, page } = req.query;
-    const p = parseInt((page as string) ?? 1);
+    let p = Number(page);
+
+    if (isNaN(p) || p < 1) {
+      p = 1;
+    }
+
     const offset = (p - 1) * perPage;
 
-    console.log(userId, q, perPage, offset);
+    let searchQuery = typeof q === "string" ? q.trim() : "";
+    if (searchQuery.length > 100) {
+      return res
+        .status(400)
+        .json({ message: "La query di ricerca è troppo lunga" });
+    }
 
     const friends = Friendship.findFriendishipByName(
       userId,
       "ACCEPTED",
-      String(q),
+      searchQuery,
       perPage,
       offset,
     );
-    console.log(friends);
+
     return res.status(200).json({ data: friends });
   } catch (error) {
     console.error(error);
@@ -77,14 +87,25 @@ export const searchUsers = (req: Request, res: Response) => {
 
   try {
     const { q, page } = req.query;
-    const p = parseInt((page as string) ?? 1);
+    let p = Number(page);
+
+    if (isNaN(p) || p < 1) {
+      p = 1;
+    }
+
     const offset = (p - 1) * perPage;
 
-    console.log(q, page);
+    let searchQuery = typeof q === "string" ? q.trim() : "";
+
+    if (searchQuery.length > 100) {
+      return res
+        .status(400)
+        .json({ message: "La query di ricerca è troppo lunga" });
+    }
 
     const foundUsers = User.searchByUsername(
       userId,
-      String(q),
+      searchQuery,
       perPage,
       offset,
     );
@@ -110,15 +131,44 @@ export const searchUsers = (req: Request, res: Response) => {
 };
 
 export const addFriendRequest = (req: Request, res: Response) => {
-  console.log("ADD FRIEND");
   const userId = res.locals.userId;
   const { receiverUserId } = req.body;
 
+  if (!receiverUserId) {
+    return res.status(400).json({ message: "ID destinatario mancante" });
+  }
+
+  if (userId === Number(receiverUserId)) {
+    return res
+      .status(400)
+      .json({ message: "Non puoi inviare una richiesta a te stesso" });
+  }
+
   //* in friendship table, userId1 has to be always the smallest between the two users
-  const userId1 = Math.min(userId, receiverUserId);
-  const userId2 = Math.max(userId, receiverUserId);
+  const userId1 = Math.min(userId, Number(receiverUserId));
+  const userId2 = Math.max(userId, Number(receiverUserId));
 
   try {
+    const existingFriendship = Friendship.findFriendshipByUsers(
+      userId1,
+      userId2,
+    );
+
+    if (existingFriendship) {
+      switch (existingFriendship.status) {
+        case "PENDING":
+          return res
+            .status(400)
+            .json({ message: "C'è già una richiesta in sospeso tra voi" });
+
+        case "ACCEPTED":
+          return res.status(400).json({ message: "Siete già amici" });
+
+        default:
+          return res.status(400).json({ message: "Relazione già esistente" });
+      }
+    }
+
     Friendship.addFriend(userId1, userId2, "PENDING", userId);
 
     return res.status(200).json({ message: "Sent Request" });
@@ -144,6 +194,30 @@ export const acceptFriendRequest = (req: Request, res: Response) => {
   const userId2 = Math.max(userId, Number(senderUserId));
 
   try {
+    const existingFriendship = Friendship.findFriendshipByUsers(
+      userId1,
+      userId2,
+    );
+
+    if (!existingFriendship) {
+      return res
+        .status(404)
+        .json({ message: "Nessuna richiesta di amicizia trovata" });
+    }
+
+    if (existingFriendship.status !== "PENDING") {
+      return res.status(400).json({
+        message: "Impossibile accettare",
+      });
+    }
+
+    if (existingFriendship.senderUserId === userId) {
+      return res.status(403).json({
+        message:
+          "Non puoi accettare una richiesta di amicizia inviata da te stesso!",
+      });
+    }
+
     Friendship.updateFriend(userId1, userId2, "ACCEPTED");
 
     return res.status(200).json({ message: "Accepted Request" });
@@ -160,11 +234,32 @@ export const deleteFriendRequest = (req: Request, res: Response) => {
   const userId = res.locals.userId;
   const { senderUserId } = req.params;
 
+  if (!senderUserId) {
+    return res.status(400).json({ message: "Missing Params" });
+  }
+
   //* in friendship table, userId1 has to be always the smallest between the two users
   const userId1 = Math.min(userId, Number(senderUserId));
   const userId2 = Math.max(userId, Number(senderUserId));
 
   try {
+    const existingFriendship = Friendship.findFriendshipByUsers(
+      userId1,
+      userId2,
+    );
+
+    if (!existingFriendship) {
+      return res
+        .status(404)
+        .json({ message: "Nessuna richiesta di amicizia trovata" });
+    }
+
+    if (existingFriendship.status !== "PENDING") {
+      return res.status(400).json({
+        message: "Impossibile eliminare",
+      });
+    }
+
     Friendship.deleteFriend(userId1, userId2, "PENDING");
 
     return res.status(200).json({ message: "Declined Request" });
@@ -190,6 +285,23 @@ export const removeFriend = (req: Request, res: Response) => {
   const userId2 = Math.max(userId, Number(friendId));
 
   try {
+    const existingFriendship = Friendship.findFriendshipByUsers(
+      userId1,
+      userId2,
+    );
+
+    if (!existingFriendship) {
+      return res.status(404).json({
+        message: "Nessun legame di amicizia trovato con questo utente",
+      });
+    }
+
+    if (existingFriendship.status !== "ACCEPTED") {
+      return res.status(400).json({
+        message: "Impossibile rimuovere l'amico",
+      });
+    }
+
     Friendship.deleteFriend(userId1, userId2, "ACCEPTED");
     return res.status(200).json({ message: "Removed friend" });
   } catch (error) {
